backend/requirements.txt
# Python 3.13 Compatible Versions
fastapi==0.115.0
uvicorn[standard]==0.32.0
supabase==2.9.0
pydantic==2.12.2
pydantic-settings==2.6.0
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
python-multipart==0.0.12
python-dotenv==1.0.1
email-validator==2.2.0
